Add human readable contributor terms, take two
Add another plain English version of the contributor terms, but this
time use the OpenStreetMap Foundation version:

    http://www.osmfoundation.org/wiki/License/Contributor_Terms_Summary

It's exactly like theirs, aside from a slight change of wording in the
opening paragraph to accommodate the layout of /user/terms, I've also
added links to the full terms, the CC-BY-SA, and the ODbL.

The new layout of the terms has each bullet point in a separate
message in en.yml. This should make things easier for the translators
than the HTML dump in the old version. I'm trusting that they won't
have to reword it in a way that stretches past these bullet points,
since each point is a summary of the equivalent contributor term
summary that's a fair assumption.

I also didn't squash this commit with "Add human readable contributor
terms" to preserve history.